+++ README.md
@@ -147,47 +147,48 @@ import time
 
 
 def response_text(openai_resp):
-    return openai_resp['choices'][0]['message']['content']
+  return openai_resp['choices'][0]['message']['content']
+
 
 from gptcache.core import cache
 from gptcache.adapter import openai
 from gptcache.embedding import Towhee
 from gptcache.cache.factory import get_ss_data_manager
-from gptcache.similarity_evaluation.simple import SearchDistanceEvaluation
+from gptcache.similarity_evaluation.distance import SearchDistanceEvaluation
 
 print("Cache loading.....")
 
 towhee = Towhee()
 data_manager = get_ss_data_manager("sqlite", "faiss", dimension=towhee.dimension())
 cache.init(
-    embedding_func=towhee.to_embeddings,
-    data_manager=data_manager,
-    similarity_evaluation=SearchDistanceEvaluation(),
-    )
+  embedding_func=towhee.to_embeddings,
+  data_manager=data_manager,
+  similarity_evaluation=SearchDistanceEvaluation(),
+)
 cache.set_openai_key()
 
 questions = [
-    "what's github",
-    "can you explain what GitHub is",
-    "can you tell me more about GitHub"
-    "what is the purpose of GitHub"
+  "what's github",
+  "can you explain what GitHub is",
+  "can you tell me more about GitHub"
+  "what is the purpose of GitHub"
 ]
 
 for question in questions:
-    for _ in range(2):
-        start_time = time.time()
-        response = openai.ChatCompletion.create(
-            model='gpt-3.5-turbo',
-            messages=[
-                {
-                    'role': 'user',
-                    'content': question
-                }
-            ],
-        )
-        print(f'Question: {question}')
-        print("Time consuming: {:.2f}s".format(time.time() - start_time))
-        print(f'Answer: {response_text(response)}\n')
+  for _ in range(2):
+    start_time = time.time()
+    response = openai.ChatCompletion.create(
+      model='gpt-3.5-turbo',
+      messages=[
+        {
+          'role': 'user',
+          'content': question
+        }
+      ],
+    )
+    print(f'Question: {question}')
+    print("Time consuming: {:.2f}s".format(time.time() - start_time))
+    print(f'Answer: {response_text(response)}\n')
 ```
 
 </details>
