+++ examples/mysql_milvus_mock/mysql_milvus_mock.py
@@ -1,7 +1,7 @@
 from gptcache.adapter import openai
 from gptcache.core import cache, Config
 from gptcache.cache.factory import get_ss_data_manager
-from gptcache.similarity_evaluation.simple import SearchDistanceEvaluation
+from gptcache.similarity_evaluation.distance import SearchDistanceEvaluation
 import numpy as np
