+++ examples/sqlite_chromadb_mock/sqlite_chromadb_mock.py
@@ -4,7 +4,7 @@ from chromadb.config import Settings
 from gptcache.adapter import openai
 from gptcache.core import cache, Config
 from gptcache.cache.factory import get_ss_data_manager
-from gptcache.similarity_evaluation.simple import SearchDistanceEvaluation
+from gptcache.similarity_evaluation.distance import SearchDistanceEvaluation
 import numpy as np
