+++ examples/sqlite_faiss_towhee/sqlite_faiss_towhee.py
@@ -4,7 +4,7 @@ import time
 from gptcache.adapter import openai
 from gptcache.core import cache, Config
 from gptcache.cache.factory import get_ss_data_manager
-from gptcache.similarity_evaluation.simple import SearchDistanceEvaluation
+from gptcache.similarity_evaluation.distance import SearchDistanceEvaluation
 from gptcache.embedding import Towhee
