+++ gptcache/similarity_evaluation/exact_match.py
@@ -1,7 +1,7 @@
 from .similarity_evaluation import SimilarityEvaluation
 
 
-class AbsoluteEvaluation(SimilarityEvaluation):
+class ExactMatchEvaluation(SimilarityEvaluation):
 
     def evaluation(self, src_dict, cache_dict, **kwargs):
         return 1 if cache_dict["question"] == src_dict["question"] else 0
