+++ tests/integration_tests/test_sqlite_faiss_towhee.py
@@ -9,7 +9,7 @@ from gptcache.adapter import openai
 from gptcache.cache.factory import get_ss_data_manager
 from gptcache.core import cache, Config
 from gptcache.embedding import Towhee
-from gptcache.similarity_evaluation.simple import SearchDistanceEvaluation
+from gptcache.similarity_evaluation.distance import SearchDistanceEvaluation
 
 
 class TestSqliteInvalid(Base):
